// src/app/models/profil.ts
export interface Profil {
    id: number;
    firstName: string;
    lastName: string; // Utilisez une chaîne de caractères ou une date, en fonction de la manière dont vous souhaitez gérer les dates
    phoneNumber: string;
    userType: string;
    siren: string;
    facebook: string;
    instagram: string;
    twitter: string;
    email: string;
    tiktok: string;
    youtube: string;
}